角色服务

+++ mingdao/mingdao-manage-dao/src/main/java/com/mingdao/dao/base/IBaseDao.java
@@ -6,9 +6,9 @@ import java.util.Map;
 import com.github.miemiedev.mybatis.paginator.domain.PageBounds;
 
 public interface IBaseDao<E> {
-	public int insertVO(E vo);
+	public Long insertVO(E vo);
 	public int insertVOs(E[] vos);
-	public int updateVO(E vo);
+	public Long updateVO(E vo);
 	public void deleteByIds(String[] ids);
 	public E queryById(String id);
 	public E queryByCondition(Map<String,Object> param);
